pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "todoist_sync"
version = "0.1.0"
description = "Some scripts to sync my needs with todoist"
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

# Dependency definitions for your project
[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.28.0"
python-dotenv = "^0.21.0"
returns = "^0.22.0"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"# If your package is meant to provide command-line executables, configure them like so:

[tool.poetry.scripts]
start = "todoist_sync.main:main"  # Adjust the module path and function as needed

# Optional dependencies can be configured as extras for optional features
[tool.poetry.extras]
gui = ["PyQt5"]
cli = ["click", "rich"]

[tool.poetry.urls]
Homepage = "https://example.com"
Repository = "https://github.com/yourusername/my-project.git"
Documentation = "https://readthedocs.org/projects/my-project/"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
mypy = "^1.9.0"

[tool.mypy]
plugins = ["returns.contrib.mypy.returns_plugin"]
